Reconcile BitScore Solidity implementation to the approved 0-100 scale
BitScoreManager.sol and IBitScoreManager.sol previously still ran on
the original 0-1000 design even after docs/bitscore-specification.md
was rescaled to 0-100. This closes that gap:

- getScore now returns uint8 on the 0-100 scale (start 30, tiers at
  25/50/75, max positive contribution +70, liquidation penalties
  -10/-5, permanent bad-debt penalty -30), matching the approved
  specification's formula exactly.
- positiveContribution/tenureCredited narrowed to uint8;
  liquidationPenalty/badDebtPenalty deliberately kept uint16 since
  they're unbounded downward accumulators.
- BitScoreManager.t.sol and BitVInvariant.t.sol's BitScore invariants
  updated off every old-scale literal; _repayUntilTier's iteration
  budgets recalibrated empirically for the new per-event point deltas.
- Repo-wide sweep for 1000/300/0-1000 literals confirms no production
  BitScore logic remains on the old scale; remaining references are
  historical documentation or explanatory comments only.
- Full Foundry suite actually executed post-change: 6 suites, 71
  tests, 71 passed, 0 failed (including the fuzzed invariant suite),
  confirming the pre-existing lending/pool/compliance/liquidation
  tests still pass unchanged.
- docs/bitscore-implementation.md rewritten to describe the new 0-100
  implementation; docs/development-log.md gets a new milestone entry.

+++ contracts/src/interfaces/IBitScoreManager.sol
@@ -10,7 +10,9 @@ pragma solidity ^0.8.24;
  * docs/bitscore-implementation.md.
  */
 interface IBitScoreManager {
-    function getScore(address user) external view returns (uint16 score);
+    /// @notice Score on BitV's 0-100 scale (docs/bitscore-specification.md),
+    /// not the earlier 0-1000 scale — uint8 is sufficient range.
+    function getScore(address user) external view returns (uint8 score);
     function getTier(address user) external view returns (uint8 tier);
 
     /// @notice Returns the score-adjusted available-borrow-value ceiling
